+++ src/mpd_client/mpd_client_settings.h
@@ -9,5 +9,5 @@
 bool mpd_api_settings_set(t_config *config, t_mpd_client_state *mpd_client_state, struct json_token *key, 
                           struct json_token *val, bool *mpd_host_changed, bool *jukebox_changed,
                           bool *check_mpd_error);
-sds mpd_client_put_settings(t_mpd_client_state *mpd_client_state, sds buffer, sds method, int request_id);
+sds mpd_client_put_settings(t_mpd_client_state *mpd_client_state, sds buffer, sds method, long request_id);
 #endif
