+++ src/mympd_api/mympd_api_bookmarks.h
@@ -8,7 +8,7 @@
 #define __MYMPD_API_BOOKMARKS_H
 bool mympd_api_bookmark_update(t_config *config, const int id, const char *name,
                                const char *uri, const char *type);
-sds mympd_api_bookmark_list(t_config *config, sds buffer, sds method, int request_id, 
+sds mympd_api_bookmark_list(t_config *config, sds buffer, sds method, long request_id, 
                             unsigned int offset);
 bool mympd_api_bookmark_clear(t_config *config);
 #endif
