+++ src/mympd_api/mympd_api_syscmds.h
@@ -6,6 +6,6 @@
 
 #ifndef __MYMPD_API_SYSCMDS_H
 #define __MYMPD_API_SYSCMDS_H
-sds mympd_api_syscmd(t_config *config, sds buffer, sds method, int request_id, 
+sds mympd_api_syscmd(t_config *config, sds buffer, sds method, long request_id, 
                      const char *cmd);
 #endif
