+++ src/tiny_queue.h
@@ -9,7 +9,7 @@
 
 typedef struct tiny_msg_t {
     void *data;
-    int id;
+    long id;
     time_t timestamp;
     struct tiny_msg_t *next;
 } tiny_msg_t;
@@ -24,8 +24,8 @@ typedef struct tiny_queue_t {
 
 tiny_queue_t *tiny_queue_create(void);
 void tiny_queue_free(tiny_queue_t *queue);
-int tiny_queue_push(struct tiny_queue_t *queue, void *data, int id);
-void *tiny_queue_shift(struct tiny_queue_t *queue, int timeout, int id);
+int tiny_queue_push(struct tiny_queue_t *queue, void *data, long id);
+void *tiny_queue_shift(struct tiny_queue_t *queue, int timeout, long id);
 void *tiny_queue_expire(tiny_queue_t *queue, time_t max_age);
 unsigned tiny_queue_length(struct tiny_queue_t *queue, int timeout);
 #endif
